Move focuszone e2etest component to diff file
